Day 16 Commit
Here is my commit for Day 16. There were 7 tutorials to follow and I followed along in my code, but at the end of the day 16 tutorials, the instructor said to return the code back to its original form. Therefore, this commit is essentially just a clean slate of starting code.