Day 20 Commit
Here is my commit for Day 20. There were 6 topics to follow and I followed along in my code, but at the end of the day 20 tutorials, the instructor said to return the code back to its original form. Therefore, this commit is essentially just a clean slate of starting code.