Day 28 Commit
This is my commit for the last challenge of day 28, which is to change the user interface so that the recommended bedtime would always appear on the screen without the need for a button. This was pretty challenging on its own, but I went the extra mile after solving it. I also made the custom views reproducible to reduce the amount of times I would code "onChange" if I were to expand on this app.